+++ src/main/java/com/manchui/domain/dto/gathering/GatheringCreateRequest.java
@@ -1,8 +1,8 @@
 package com.manchui.domain.dto.gathering;
 
-import com.manchui.domain.entity.ChatRoom;
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
 import com.manchui.domain.entity.Gathering;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
