+++ src/main/java/com/manchui/domain/dto/gathering/GatheringCreateResponse.java
@@ -1,6 +1,5 @@
 package com.manchui.domain.dto.gathering;
 
-import com.manchui.domain.entity.ChatRoom;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
