+++ src/main/java/com/manchui/domain/entity/Gathering.java
@@ -1,7 +1,9 @@
 package com.manchui.domain.entity;
 
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
 import com.manchui.domain.dto.gathering.ClosedGatheringInfoResponse;
 import com.manchui.domain.dto.gathering.GatheringCreateResponse;
+import com.manchui.domain.user.entity.User;
 import com.manchui.global.entity.Timestamped;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Max;
