+++ src/main/java/com/manchui/domain/entity/Review.java
@@ -2,6 +2,7 @@ package com.manchui.domain.entity;
 
 import com.manchui.domain.dto.review.ReviewCreateRequest;
 import com.manchui.domain.dto.review.ReviewCreateResponse;
+import com.manchui.domain.user.entity.User;
 import com.manchui.global.entity.Timestamped;
 import jakarta.persistence.*;
 import lombok.*;
