+++ src/main/java/com/manchui/domain/notification/service/NotificationServiceImpl.java
@@ -3,7 +3,7 @@ package com.manchui.domain.notification.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 import com.manchui.domain.notification.dto.NotificationCursorPagingResponse;
 import com.manchui.domain.notification.dto.NotificationResponse;
 import com.manchui.domain.notification.entity.Notification;
@@ -11,7 +11,7 @@ import com.manchui.domain.notification.entity.NotificationContent;
 import com.manchui.domain.notification.entity.NotificationType;
 import com.manchui.domain.notification.repository.EmitterRepository;
 import com.manchui.domain.notification.repository.NotificationRepository;
-import com.manchui.domain.service.UserService;
+import com.manchui.domain.user.service.UserService;
 import com.manchui.global.exception.CustomException;
 import com.manchui.global.exception.ErrorCode;
 import com.manchui.global.response.SuccessResponse;
