+++ src/main/java/com/manchui/domain/repository/HeartRepository.java
@@ -2,7 +2,7 @@ package com.manchui.domain.repository;
 
 import com.manchui.domain.entity.Gathering;
 import com.manchui.domain.entity.Heart;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
