+++ src/main/java/com/manchui/domain/repository/ReviewRepository.java
@@ -2,7 +2,7 @@ package com.manchui.domain.repository;
 
 import com.manchui.domain.entity.Gathering;
 import com.manchui.domain.entity.Review;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 import com.manchui.domain.repository.querydsl.ReviewQueryDsl;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
