+++ src/main/java/com/manchui/domain/service/GatheringStore.java
@@ -1,9 +1,9 @@
 package com.manchui.domain.service;
 
 import com.manchui.domain.dto.gathering.GatheringCreateRequest;
-import com.manchui.domain.entity.ChatRoom;
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
 import com.manchui.domain.entity.Gathering;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 
 import java.time.LocalDateTime;
