+++ src/main/java/com/manchui/domain/service/GatheringStoreImpl.java
@@ -2,9 +2,9 @@ package com.manchui.domain.service;
 
 import com.manchui.domain.dto.gathering.GatheringCreateRequest;
 import com.manchui.domain.entity.Attendance;
-import com.manchui.domain.entity.ChatRoom;
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
 import com.manchui.domain.entity.Gathering;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 import com.manchui.domain.repository.AttendanceRepository;
 import com.manchui.domain.repository.GatheringRepository;
 import jakarta.transaction.Transactional;
