+++ src/main/java/com/manchui/domain/service/ReviewServiceImpl.java
@@ -4,9 +4,10 @@ import com.manchui.domain.dto.review.*;
 import com.manchui.domain.entity.Attendance;
 import com.manchui.domain.entity.Gathering;
 import com.manchui.domain.entity.Review;
-import com.manchui.domain.entity.User;
+import com.manchui.domain.user.entity.User;
 import com.manchui.domain.repository.AttendanceRepository;
 import com.manchui.domain.repository.ReviewRepository;
+import com.manchui.domain.user.service.UserService;
 import com.manchui.global.exception.CustomException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
